// resolvers/src/models/API.ts
export type BookInput = {
  title?: string | null,
  publicationDate?: string | null,
  authorId?: string | null,
  genreIds?: Array< string | null > | null,
};

export type Book = {
  __typename: "Book",
  bookId?: string | null,
  title?: string | null,
  publicationDate?: string | null,
  authorId?: string | null,
};
// ...
export type PutBookMutationVariables = {
  bookId: string,
  input: BookInput,
};

export type PutBookMutation = {
  putBook?:  {
    __typename: "Book",
    bookId?: string | null,
    title?: string | null,
    publicationDate?: string | null,
    authorId?: string | null,
  } | null,
};